MDEV-30705: JSON_SCHEMA_VALID: schema with multipleOf for big value always return 1
Analysis: Implementation used double to store value. longlong is better
choice
Fix: Use longlong for multiple_of and the value to store it and num_flag to
check for decimals.

+++ sql/json_schema.h
@@ -150,7 +150,7 @@ class Json_schema_minimum : public Json_schema_keyword
 class Json_schema_multiple_of : public Json_schema_keyword
 {
   private:
-    double multiple_of;
+    longlong multiple_of;
 
   public:
     bool validate(const json_engine_t *je) override;
